Revert "PatchClampSeries: Force sweep_number to uint64"
This reverts commit 6b6033e2044146d6d2448f278985c419d7ea4c4b.

+++ src/pynwb/icephys.py
@@ -1,5 +1,3 @@
-import numpy as np
-
 from hdmf.utils import docval, popargs, call_docval_func, get_docval
 
 from . import register_class, CORE_NAMESPACE
@@ -90,7 +88,7 @@ class PatchClampSeries(TimeSeries):
             if not (sweep_number >= 0):
                 raise ValueError("sweep_number must be a non-negative integer")
 
-            self.sweep_number = np.uint64(sweep_number)
+            self.sweep_number = sweep_number
 
 
 @register_class('CurrentClampSeries', CORE_NAMESPACE)
